Corrected bugs

+++ src/map/include/entitysystem.h
@@ -130,7 +130,7 @@ class EntitySystem {
   std::vector<std::unique_ptr<CommandBase>> delete_commands_;
 
  private:
-  static constexpr double drop_radius = 50.f;
+  static constexpr float drop_radius = 50.f;
 };
 
 #endif /* !_ENTITYSYSTEM_H_ */
